pkg/config: move config types from utils
