+++ pkg/utils/duration.go
@@ -1,93 +1,16 @@
 package utils
 
 import (
-	"database/sql/driver"
-	"encoding/json"
-	"fmt"
 	"time"
 
-	"github.com/pkg/errors"
+	"github.com/smartcontractkit/chainlink-relay/pkg/config"
 )
 
-// Duration is a non-negative time duration.
-type Duration struct{ d time.Duration }
+// Deprecated: use config.Duration
+type Duration = config.Duration
 
-func NewDuration(d time.Duration) (Duration, error) {
-	if d < time.Duration(0) {
-		return Duration{}, fmt.Errorf("cannot make negative time duration: %s", d)
-	}
-	return Duration{d: d}, nil
-}
+// Deprecated: use config.NewDuration
+func NewDuration(d time.Duration) (Duration, error) { return config.NewDuration(d) }
 
-func MustNewDuration(d time.Duration) *Duration {
-	rv, err := NewDuration(d)
-	if err != nil {
-		panic(err)
-	}
-	return &rv
-}
-
-func (d Duration) Duration() time.Duration {
-	return d.d
-}
-
-func (d Duration) String() string {
-	return d.d.String()
-}
-
-// MarshalJSON implements the json.Marshaler interface.
-func (d Duration) MarshalJSON() ([]byte, error) {
-	return json.Marshal(d.String())
-}
-
-// UnmarshalJSON implements the json.Unmarshaler interface.
-func (d *Duration) UnmarshalJSON(input []byte) error {
-	var txt string
-	err := json.Unmarshal(input, &txt)
-	if err != nil {
-		return err
-	}
-	v, err := time.ParseDuration(string(txt))
-	if err != nil {
-		return err
-	}
-	*d, err = NewDuration(v)
-	if err != nil {
-		return err
-	}
-	return nil
-}
-
-// MarshalText implements the text.Marshaler interface.
-func (d Duration) MarshalText() ([]byte, error) {
-	return []byte(d.d.String()), nil
-}
-
-// UnmarshalText implements the text.Unmarshaler interface.
-func (d *Duration) UnmarshalText(input []byte) error {
-	v, err := time.ParseDuration(string(input))
-	if err != nil {
-		return err
-	}
-	pd, err := NewDuration(v)
-	if err != nil {
-		return err
-	}
-	*d = pd
-	return nil
-}
-
-func (d *Duration) Scan(v interface{}) (err error) {
-	switch tv := v.(type) {
-	case int64:
-		*d, err = NewDuration(time.Duration(tv))
-		return err
-	default:
-		return errors.Errorf(`don't know how to parse "%s" of type %T as a `+
-			`models.Duration`, tv, tv)
-	}
-}
-
-func (d Duration) Value() (driver.Value, error) {
-	return int64(d.d), nil
-}
+// Deprecated: use config.MustNewDuration
+func MustNewDuration(d time.Duration) *Duration { return config.MustNewDuration(d) }
